Fold NettyRequestBodyWithHttpDataFactory into NettyRequestBody
Every NettyRequestBody implementation (future/cats/zio via NettyMonadRequestBody,
and sync) uses the multipart httpDataFactory, so the intermediate trait added no
real separation. Move the factory and its multipartTempDirectory/multipartMinSizeForDisk
config into the base trait and drop NettyRequestBodyWithHttpDataFactory, flattening
the hierarchy by one level.
